requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
transformers==4.57.3
torch==2.9.1
lingua-language-detector==2.1.1
sentencepiece  # Required for NLLB usually, though not in freeze output maybe installed implicitly or system. Adding to be safe.
protobuf # Often needed by transformers
requests==2.32.5
pytest
httpx
